% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicodeTable.R
\name{unicodeTable}
\alias{unicodeTable}
\title{Function to print unicode table from a dataframe}
\usage{
unicodeTable(data, addRowNames = TRUE)
}
\arguments{
\item{data}{A data.frame or tibble to print}

\item{addRowNames}{Logical. Whether to add a 0-based index column.}
}
\value{
prints the table to console.
}
\description{
pass a dataframe as input the function will print to console a unicode table
}
\examples{
sep_line <- "──────────────────────────"
sep_score <- "─────────"
sep_total <- "──────────"

data <- data.frame(
   `Metric Assessed` = c(
     "Statistical methodology", 
     "Replicability", 
     "Conclusions", 
     "Presentation", 
     "Citations",
     sep_line,          
     "Overall",
     sep_line    
   ),
   `Score` = c("17", "5", "5", "3", "2", sep_score, "32", sep_score),
   `Total Marks` = c("20", "5", "5", "5", "5", sep_total, "40", sep_total),
   check.names = FALSE
)
unicodeTable(data)
╭───┬────────────────────────────┬───────────┬─────────────╮ 
│   │ Metric Assessed            │ Score     │ Total Marks │ 
├───┼────────────────────────────┼───────────┼─────────────┤ 
│ 0 │ Statistical methodology    │ 17        │ 20          │ 
│ 1 │ Replicability              │ 5         │ 5           │ 
│ 2 │ Conclusions                │ 5         │ 5           │ 
│ 3 │ Presentation               │ 3         │ 5           │ 
│ 4 │ Citations                  │ 2         │ 5           │ 
│ 5 │────────────────────────────│───────────│─────────────│ 
│ 6 │ Overall                    │ 32        │ 40          │ 
│ 7 │────────────────────────────│───────────│─────────────│ 
╰───┴────────────────────────────┴───────────┴─────────────╯ 
}
